added base npc class
