+++ src/sprites/elf/index.ts
@@ -1,4 +1,4 @@
-export * from "./hub/teleporter"
+export * from "./minigame/teleporter"
 export * from "./minigame/npc"
 export * from "./minigame/player"
 export * from "./minigame/fruits/apple"
