+++ src/sprites/game/index.ts
@@ -1,2 +1,1 @@
-export * from "./npc"
 export * from "./player"
