+++ src/sprites/goblin/index.ts
@@ -1,3 +1,3 @@
-export * from "./hub/teleporter"
+export * from "./minigame/teleporter"
 export * from "./minigame/npc"
 export * from "./minigame/objective"
