+++ src/sprites/index.ts
@@ -1,3 +1,4 @@
 export * from "./input"
 export * from "./dialogue"
 export * from "./spritesUtilities"
+export * from "./npc"
